requirements.txt
discord.py
python-dotenv
aiohttp
replicate
Pillow
openai
pytest
